Split ServersInfoExtractor class in an interface+ObjectTreeReader class
